+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Date & Domain Helpers
 
 Presentation
 
+- `PaddedUIView` is added, that pads underlying `UIView`
 - `appearanceDelay` is added to `ProgressViewParameters`
 
 ### [8.1.0(114)](https://github.com/VakhoKontridze/VCore/releases/tag/8.1.0) — 2026 05 04